fix(deploy): skip husky when dev dependencies omitted
